lang: move engine/util funcs into lang package
This avoids circular import references in the lang/types package when
calling these functions from lang/types. It could be argued that these
never really belonged in the engine/util package to begin with.

This is almost a direct code move, except for a few functions that have
been renamed for brevity and the function signature adjusted for
reusability particularly with regards to passing reflect.Type instead of
a concrete type, which makes the function much more useful in the
general case.
